+++ next.config.ts
@@ -19,7 +19,20 @@ const nextConfig: NextConfig = {
     ]
   },
   poweredByHeader: false,
-  reactStrictMode: true
+  reactStrictMode: true,
+  async headers() {
+    return [
+      {
+        headers: [
+          { key: "Referrer-Policy", value: "strict-origin-when-cross-origin" },
+          { key: "Strict-Transport-Security", value: "max-age=31536000" },
+          { key: "X-Content-Type-Options", value: "nosniff" },
+          { key: "X-Frame-Options", value: "DENY" }
+        ],
+        source: "/:path*"
+      }
+    ];
+  }
 };
 
 export default nextConfig;